Add pip-compile'd requirements.txt
In theory, this file shouldn't need to be regenerated very often.
One thing of note is that pip-compile strongly recommends that it be run
on the target system in order to ensure it selects the correct versions.
This is because pip-compile doesn't (yet) have a way to specify the
Python version, platform, and machine type during compilation.
The setup.py file was also tweaked to ignore any --options in the
requirements.txt file.